+++ components/dash-html-components/tests/test_integration.py
@@ -114,8 +114,8 @@ def test_click_static(dash_duo):
         [
             html.Div("no event listener", className="div-1"),
             html.Div("event listener", id="div-2", n_clicks=0),
-            html.Div("no event listener", id="div-3", n_clicks=0, static=True),
-            html.Div("event listener", id="div-4", n_clicks=0, static=False),
+            html.Div("no event listener", id="div-3", n_clicks=0, disable_n_clicks=True),
+            html.Div("event listener", id="div-4", n_clicks=0, disable_n_clicks=False),
             html.Div(id="div-output"),
         ]
     )